% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infection_treatment.R
\docType{data}
\name{infection_treatment}
\alias{infection_treatment}
\title{Infection and Treatments}
\format{
A data frame with 150 rows and 2 variables:
\describe{
  \item{Infection}{Indicates whether the person was diagnosed with an infection.}
  \item{Treatment}{Treatment option utilized.}
}
}
\source{
Original "infection_treatment.xlsx" from textbook materials.
}
\usage{
infection_treatment
}
\description{
Raw dataset containing treatments and infection outcomes.
Used in the Categorical Analysis chapter of the textbook.
}
\keyword{datasets}
