% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pizza_sales.R
\docType{data}
\name{pizza_sales}
\alias{pizza_sales}
\title{Pizza Sales (Clean)}
\format{
A data frame with 156 rows and 3 variables:
\describe{
  \item{price}{Price of pizza in USD}
  \item{volume}{Number of pizzas sold}
}
}
\source{
Derived from \code{pizza_sales_raw} using cleaning steps in linear regression chapter
}
\usage{
pizza_sales
}
\description{
Cleaned dataset containing pizza prices and sales volume prior.
Used in Linear Regression Chapter of textbook to demonstrate modeling.
}
\keyword{datasets}
