% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pizza_sales_raw.R
\docType{data}
\name{pizza_sales_raw}
\alias{pizza_sales_raw}
\title{Pizza Sales (Raw)}
\format{
A data frame with 156 rows and 3 variables:
\describe{
  \item{Week}{Week number of sales (not used in final analysis)}
  \item{Total Price}{Price of pizza in USD}
  \item{Total Volume}{Number of pizzas sold}
}
}
\source{
Original “pizza_sales.xlsx” from textbook materials.
}
\usage{
pizza_sales_raw
}
\description{
Raw dataset containing pizza prices and sales volume prior to cleaning.
Used in Linear Regression Chapter of textbook to demonstrate cleaning.
}
\keyword{datasets}
